requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
celery==5.4.0
redis==5.1.1
psycopg2-binary==2.9.9
django-celery-beat==2.7.0
flower==2.0.1
